Allows removal of Users

+++ src/pages/EditOrganization.tsx
@@ -125,8 +125,7 @@ const UsersTable = () => {
 
   const onClickRemoveUser = (removedUser:OrganizationUser) =>
   {
-    OrganizationService.deleteOrganizationUser(removedUser.organizationId, removedUser.userId)
-    .then(() => setOrgUsers(orgUsers.filter((orgUser) => orgUser.userId !== removedUser.userId)));
+    OrganizationService.deleteOrganizationUser(removedUser.organizationId, removedUser.userId).then(() => setOrgUsers(orgUsers.filter((orgUser) => orgUser.userId !== removedUser.userId)));
   }
 
   return (
